+++ users/users.mutations.js
@@ -1,0 +1,34 @@
+import client from "../client";
+import bcrypt from "bcrypt";
+
+export default {
+  Mutation: {
+    createAccount: async (
+      _,
+      { username, firstName, email, password, lastName }
+    ) => {
+      try {
+        const existingUser = await client.user.findFirst({
+          where: {
+            OR: [{ username }, { email }],
+          },
+        });
+        if (existingUser) {
+          throw new Error("This username/email is already taken.");
+        }
+        const hashedPassword = await bcrypt.hash(password, 10);
+        return client.user.create({
+          data: {
+            firstName,
+            lastName,
+            username,
+            email,
+            password: hashedPassword,
+          },
+        });
+      } catch (e) {
+        return e;
+      }
+    },
+  },
+};